multitask_ti_classification/requirements.txt
# Core ML libraries
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0
torch-spline-conv>=1.2.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Dimensionality reduction and visualization
umap-learn>=0.5.0
tsne-torch>=1.0.0

# Hyperparameter optimization
optuna>=3.0.0
ray[tune]>=2.0.0

# Progress bars and utilities
tqdm>=4.62.0
joblib>=1.1.0

# Experiment tracking (optional)
wandb>=0.12.0

# Additional utilities
h5py>=3.6.0
pyyaml>=6.0
tensorboard>=2.8.0

# For GPU support (if needed)
# cupy-cuda11x>=11.0.0  # Uncomment and adjust version for your CUDA version

# Development and testing
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0 